major changes.  See the file ~ptolemy/adm/KERNEL-CHANGES-0.4 for details.

+++ ptclassic/src/kernel/PriorityQueue.cc
@@ -20,7 +20,7 @@ $Id$
 #include "PriorityQueue.h"
 
 // Set the properties of the LevelLink class.
-LevelLink* LevelLink :: setLink(Pointer a, float v, float fv, LevelLink* n,
+LevelLink* LevelLink :: setLink(Pointer a, double v, double fv, LevelLink* n,
 			  LevelLink* b) {
 	e = a;
 	level = v;
@@ -65,7 +65,7 @@ void PriorityQueue :: clearFreeList() {
 }
 
 // constructor
-PriorityQueue :: PriorityQueue(Pointer a, float v, float fv = 1.0) :
+PriorityQueue :: PriorityQueue(Pointer a, double v, double fv = 1.0) :
 	freeLinkHead(0), numberNodes(1), numFreeLinks(0)
 {
 	lastNode = getFreeLink()->setLink(a, v, fv, 0, 0);
@@ -76,7 +76,7 @@ PriorityQueue :: PriorityQueue(Pointer a, float v, float fv = 1.0) :
 
 // lowest level first, lowest fineLevel first.
 
-LevelLink* PriorityQueue :: levelput(Pointer a, float v, float fv)
+LevelLink* PriorityQueue :: levelput(Pointer a, double v, double fv)
 {
 	numberNodes++;			// increment numberNodes
 	LevelLink* newLink = getFreeLink();
@@ -118,7 +118,7 @@ LevelLink* PriorityQueue :: levelput(Pointer a, float v, float fv)
 
 // highest level first, highest fineLevel first.
 
-LevelLink* PriorityQueue :: leveltup(Pointer a, float v, float fv)
+LevelLink* PriorityQueue :: leveltup(Pointer a, double v, double fv)
 {
 	numberNodes++;			// increment numberNodes
 	LevelLink* newLink = getFreeLink();
@@ -254,7 +254,7 @@ PriorityQueue :: ~PriorityQueue () {
 	clearFreeList();
 }
 	
-void PriorityQueue :: put (Pointer a, float v)
+void PriorityQueue :: put (Pointer a, double v)
 
 {
         // place at end of queue regardless of "number stamp"
